Add more classes for OOP
